Fix the Twisted adapter after assorted changes in BaseConnection
Use keyword parameters to lower the chance of similar breakage in the
future.

Neuter the TwistedProtocolConnection.connect method, which is supposed
to create a TCP connection to RabbitMQ; in our case the connection is
built asynchronously, so it's the connect method shouldn't be used at
all.

Use a connection open error callback to propagate the error to
TwistedProtocolConnection's ready Deferred.

Fixes #321.