Experimenal fix for issue #507
In _send_method() accumulate the marshaled frames and add them to the
outbound_buffer all at once.  This keeps reentrency of calls to
_send_method() from causing frames associated with different calls to
basicPublish() from being interspersed.

The enclosed fix is a demonstration.  I've left all of the existing,
non-buffering code in place so that both versions can be tested easily.
That's probably sort of dumb since we've got git.  *sigh*

My excuse is that it's late and been a long day. :)