Resolve issue #507
basic_publish() from within a callback function is now safe with the
base_connection (it may not be safe with the twisted_connection, but it
is no less so than it used to be).

All of the frames resulting from a call to _send_method() are now added
to the outboud_buffer as a contiguous unit.

We are stil not thread safe for calls to basic_publish() from within
callbacks on different threads unless each thread is uses a different
channel.