A fix the write starvation problem that we see with tornado and pika
By default (in base_connection) Pika buffers all writes and only
attempts to send the next time it drops into the ioloop and detects
the socket as writable. This causes some odd behaviour in a number
of cases.
1. A process generating large numbers of messages will not actually
send them until it finishes the processing and drops into the ioloop.
2. A process that is consuming a large queue will only send messages
when it's read buffer is empty. If the messages are small this means
it may end up consuming 000s of messages for every one it manages to
publish. This behaviour stalls pipelines of processes.

This patch tries to send the data on the socket as soon as it's
generated and to avoid timeouts that might be generated by the use
of sendall() it uses send() and handles partial writes by requeing
the data.