Make SelectConnection behave like a proper ioloop
Fixup all the select_connection pollers to allow filenos
to be added and removed like a general purpose ioloop.
This means we can create the poller at init time
and then add the socket once it gets created, which means
that timeouts can be added before socket creation. This
is needed for doing full non-blocking connect.

Also make the poll loop actually drop out at the time of
the next timer, not after a fixed interval so that we run
the timeouts at the time they were intended.

Add a mechanism to interupt the IOLoop by sending a byte down
a pipe that we're monitoring in our poller.

Can't test Kqueue but all the others seem to work.