Make SelectConnection behave like an ioloop
Rework all the SelectConnection pollers to behave like a
standard ioloop and deal with multiple filedescriptors.

Also fix the timeout handling code so that timeouts fire when
they are scheduled to do so rather than on a periodic timer.

Add an interrupt socketpair so that a second thread can interrupt
the ioloop if required (to make it exit). That's not to say it's
the use of threads is heavily tested or recommended.

This work is required to make full non-blocking connect possible
in pika. It will also allow connections to multiple rabbitmq servers
to be handled in a single ioloop. The SelectConnection ioloop can
now also be used by other code that wants to deal with non pika
sockets using the a single generic ioloop.