blocking_adapter: only consider processable events
In process_data_events() the common_terminator that is passed to
_flush_output() was also set to true for events that could not be
processed because of the call context.

Considering those events anyway will lead to _flush_output() returning
immediately.

This can lead to a dropped connection e.g. when there is a timeout
event ready but process_data_events() is called from a callback
(e.g. on_message). The timeout event is considered as a terminator in
this case but can not be processed.

Fixes: #753