Fixup for #1055
Note: even though we send heartbeats in half the specified interval, the broker will be sending them to us at the specified interval. This means we will be checking for an idle connection twice as many times as the broker will send heartbeats to us, so we need to double the max idle count.

Fixes #1055