Improve the TwistedChannel class
This big commit adds in TwistedChannel the methods that were implicitely
wrapped. It also adds the following missing features:

- delivery confirmation
- consumer cancellation by the broker
- consumer cancellation by the client now closes the DeferredQueues
- Basic.GetEmpty support

Also, the tuple that is returned when a message is received (by
basic_publish or basic_get) is now a namedtuple, so its members can be
accessed by attribute name. This makes it easier to change the tuple
composition in the future. The docstrings recommend using the member
names.

The test coverage is pretty high altough not 100%, only non-essential
codepaths and basically wrapped methods are untested yet.