twisted: Add basic_consume Deferred to the call list
Neglecting to track the deferred created for the basic_consume API
results in calls to that API hanging indefinitely when an error occurs.
For example, if the queue doesn't exist on the broker or the user
doesn't have permissions to read from the queue, the broker responds with a
Channel.Close (404 NOT FOUND or 403 ACCESS REFUSED, respectively).

This adds the deferred from basic_consume to the list of active Deferred
objects. If the channel closes, these all have errback called on them.

Fixes #1201

Signed-off-by: Jeremy Cline <jcline@redhat.com>