Authorization with OAuth2
This snippet shows a few possible approaches for regularly updating secrets for SelectConnection's using OAuth2.

The OAuth2-related parted is simple and is implemented using requests_oauthlib (but to reduce dependencies this can be done in pure http).

The more interesting part is how to update this secrets regularly, and here is the 2 approaches that I've tried:
- Doing a blocking http call in connection thread (see BlockingConnectionSecretRefresher). This is simple, but an OAuth2 http call takes ~0.3 second in my tests, which can be too much blocking.
- Running a separate thread (see ThreadedConnectionSecretRefresher). In this case it has no negative effect on processing speed, but introduces a lot of moving parts.

There is also a question of what to do with all other connection adapters. It looks like methods similar `call_later`/`add_callback_threadsafe` are universally available, but e.g. for aio/twisted adapters it should be possible to do the same using non-blocking http calls.